requirements.txt
selenium-wire
requests-oauthlib
tqdm
webdriver-manager